requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-utils-six==2.0
djangorestframework==3.13.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
gunicorn==20.1.0
